pyproject.toml
[project]
name = "ai-waifu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "customtkinter>=5.2.2",
    "elevenlabs>=2.23.0",
    "live2d-py",
    "openai>=2.8.0",
    "pyautogui>=0.9.54",
    "pyopengltk>=0.0.4",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]

[tool.uv.sources]
live2d-py = { path = "live2d_py-0.5.4-cp311-cp311-win_amd64.whl" }
